refactor: move AcrossApi types into separate file
